+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [Backup NG form] Add offline snapshot info (PR [#3144](https://github.com/vatesfr/xen-orchestra/pull/3144))
 - [Backup NG overview] Display concurrency and offline snapshot value [3087](https://github.com/vatesfr/xen-orchestra/issues/3087) (PR [3145](https://github.com/vatesfr/xen-orchestra/pull/3145))
 - [VM revert] notify the result of reverting a VM [3095](https://github.com/vatesfr/xen-orchestra/issues/3095) (PR [3150](https://github.com/vatesfr/xen-orchestra/pull/3150))
+- Web hooks plugin [#1946](https://github.com/vatesfr/xen-orchestra/issues/1946)
 
 ### Bug fixes
 
@@ -20,6 +21,7 @@
 
 ### Released packages
 
+- xo-server-hooks: 0.1.0
 - xo-remote-parser 0.4.0
 - @xen-orchestra/fs 0.2.0
 - vhd-lib 0.3.0